move dynamic partitions add/delete requests to new module
